% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentList.R
\name{momentList}
\alias{momentList}
\alias{mean.momentList}
\alias{cov.momentList}
\title{Class momentList}
\usage{
momentList(rawMomentOrders = NULL, rawMoments = list(),
  centralMomentOrders = NULL, centralMoments = list())

\method{mean}{momentList}(x)

\method{cov}{momentList}(x)
}
\arguments{
\item{rawMomentOrders}{matrix. Every row gives the order of a raw moment that is already known.}

\item{rawMoments}{list. The i-th entry is the raw Moment of order \code{rawMomentOrders[i, ]}.}

\item{centralMomentOrders}{matrix. Every row gives the order of a central moment that is already known.}

\item{centralMoments}{list. The i-th entry is the central Moment of order \code{centralMomentOrders[i, ]}.}

\item{x}{object of class \code{momentList}}
}
\description{
Create an object of class \code{momentList}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{mean}: pick means out of momentList

\item \code{cov}: pick covariances out of momentList
}}

