% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolicMoments.R
\name{symbolicMoments}
\alias{symbolicMoments}
\alias{symbolicmoments}
\alias{symbolicmoment}
\alias{symbolicMoment}
\title{Symbolic calculation of moments}
\usage{
symbolicMoments(distribution, missingOrders, mean = NA, cov = NA,
  var = NA, simplify = TRUE)
}
\arguments{
\item{distribution}{string specifying the (multivariate) distribution
of X. The following values are possible: 
\itemize{
\item \code{"zero"} sets all moments to 0,
\item \code{"normal"} calculates the moments of a centralized multivariate normal distribution,
\item \code{"lognormal"} calculates the raw moments of a multivariate lognormal distribution,
\item \code{"gamma"} calculates the raw moments of a multivariate gamma distribution.
}}

\item{missingOrders}{numeric vector or matrix. Each row gives the order
of a moment that shall be calculated.}

\item{mean}{vector or list with expected values. Entry \code{mean[[i]]} should
contain a value for \eqn{E(X_i)}{E(X[i])}.}

\item{cov}{matrix or nested list. Entry \code{cov[i][j]} (or \code{cov[[i]][[j]]} 
respectivly) should contain a value for the covariance \eqn{Cov(X_i, X_j)}{Cov(X[i], X[j])}.}

\item{var}{optional. If \code{n = 1}, \code{cov} would only
contain one entry - the variance \eqn{Var(X)}. This value can be passed
to parameter \code{var} instead of \code{cov}.}

\item{simplify}{bool indiciating if the resulting expressions should be simplified.
Function \code{\link[Deriv]{Simplify}} from package \pkg{Deriv} is used for simplification.}
}
\value{
A list where each element is a quoted expression.
The i-th element of this list gives a formula for the
moment whose order is given in the i-th row of \code{missingOrders}.
If \code{simplify = TRUE}, the returned value may as well be a vector or number.
}
\description{
Compute the moments of a multivariate variable 
\eqn{X = (X_1, ..., X_n)}{X = X[1], ..., X[n]}
and return the formula as quoted expression.
}
\note{
The calculation of the central moments of a multivariate \bold{normal} distribution
is based on function \code{\link[symmoments]{callmultmoments}} of package \pkg{symmoments}.
If the calculation for a  multivariate \bold{gamma} distribution leads to NaNs, 
then the values of cov and mean do not fit to a gamma distribution.
All entries of cov should be positive and the diagonals should be large
with respect to the other entries. More specifically, the inequations 
\deqn{\texttt{mean[i]} > \sum_{k \neq i} \frac{mean[k]\cdot cov[i,k]}{cov[i,i]}}{
mean[i] > sum_(k != i) mean[k]*cov[i,k]/cov[i,i]}
should be satisfied for all i in 1:n.
}
