% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformMoment.R
\name{transformMoment}
\alias{transformMoment}
\title{Transform raw moments into central moments and vice versa}
\usage{
transformMoment(order, type, momentList, closure = "zero",
  simplify = TRUE)
}
\arguments{
\item{order}{numeric vector giving the order of the desired moment}

\item{type}{string, either 'central' or 'raw'}

\item{momentList}{object of class \code{\link{momentList}}.}

\item{closure}{string giving the closure method to use if a central moment is unknown.}

\item{simplify}{bool indiciating if the resulting expressions should be simplified.
Function \code{\link[Deriv]{Simplify}} from package \pkg{Deriv} is used for simplification.}
}
\value{
This function returns an object of class 'momentList'.
}
\description{
Let X be a random variable and Y the corresponding centered variable, i.e. \eqn{Y = X - \mu}. 
Let p be the \code{order} of the desired moment. 
If \code{type = 'raw'}, this function returns
\deqn{E(X^p) = \sum_{k_1=0}^{p_1}...\sum_{k_n=0}^{p_n} {p \choose k} \mu^{p-k} E(Y^k).}
The values of \eqn{E(Y^k)} are replaced by
\itemize{
  \item \code{centralMoment[[i]]} if \code{k} = \code{centralMomentOrders[i, ]},
  \item \code{\link{transformMoment}(type = 'central', ...)} if \code{k} is a row in \code{rawMomentOrders},
  \item \code{\link{symbolicMoments}(closure, k)} if \code{k} is neither a row in \code{centralMomentOrders} nor in \code{rawMomentOrders}.
}
If \code{type = 'central'}, this function returns
\deqn{E(Y^p) = \sum_{k_1=0}^{p_1}...\sum_{k_n=0}^{p_n} (-1)^{p-k} {p \choose k} \mu^{p-k} E(X^k).}
The values of \eqn{E(X^k)} are replaced by
\itemize{
  \item \code{rawMoment[[i]]} if \code{k} = \code{rawMomentOrders[i, ]},
  \item \code{\link{transformMoment}(type = 'raw', ...)} if \code{k} is not a row in \code{rawMomentOrders}.
}
}
