% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentList.R
\name{momentList}
\alias{momentList}
\alias{mean.momentList}
\alias{cov.momentList}
\title{Class momentList}
\usage{
momentList(rawMomentOrders = NULL, rawMoments = list(),
  centralMomentOrders = NULL, centralMoments = list())

\method{mean}{momentList}(x)

\method{cov}{momentList}(x)
}
\description{
Create an object of class \code{momentList}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{mean}: pick means out of momentList

\item \code{cov}: pick covariances out of momentList
}}

